Rename ExcludedDateDisplay to DateOrIntervalText.
