+++ client/src/view/generate-tutorials/components/excluded-dates/ExcludedDateBox.tsx
@@ -2,8 +2,8 @@ import { Box, IconButton } from '@material-ui/core';
 import { createStyles, makeStyles } from '@material-ui/core/styles';
 import { Delete as DeleteIcon, SquareEditOutline as EditIcon } from 'mdi-material-ui';
 import React from 'react';
+import DateOrIntervalText from '../../../../components/DateOrIntervalText';
 import OutlinedBox from '../../../../components/OutlinedBox';
-import ExcludedDateText from './ExcludedDateDisplay';
 import { FormExcludedDate } from './FormikExcludedDates';
 
 const useStyles = makeStyles((theme) =>
@@ -25,7 +25,7 @@ function ExcludedDateBox({ excluded, onEdit, onDelete }: Props): JSX.Element {
 
   return (
     <OutlinedBox display='flex' alignItems='center'>
-      <ExcludedDateText excluded={excluded} />
+      <DateOrIntervalText date={excluded} />
 
       <Box marginLeft='auto'>
         <IconButton size='small' onClick={onEdit}>
